requirements.txt
GoogleAppEnginePipeline
GoogleAppEngineCloudStorageClient
requests
